bug fixed in Correct Transects tool
It has been identified that "Generate Transects Along Line" ArcGIS' geoprocess, which is used in the tool, may totally invert transects without an apparent reason. It thas been coded a solution to handle these situations (by adding a new function within the TransectProcessor class object).
On the other hand, it has also been corrected the situation of the transects with angles near to 360-0 range that could lead to large differences between orientations when in fact, there are no such large differences.